Clean up naming of the shared f3/f3x components
