+++ front-end/src/app/reports/f3-shared/create-shared-f3/create-shared-f3.component.scss
