+++ front-end/src/app/reports/f3-shared/f3/report-detailed-summary/report-detailed-summary.component.ts
