+++ front-end/src/app/reports/f3-shared/f3/report-summary/report-summary.component.html
