+++ front-end/src/app/reports/f3-shared/f3x/report-detailed-summary/report-detailed-summary.component.spec.ts
