+++ front-end/src/app/reports/f3-shared/f3x/report-summary/report-summary.component.scss
