+++ front-end/src/app/reports/f3-shared/f3x/routes.ts
@@ -1,16 +1,16 @@
 import { Route } from '@angular/router';
-import { ReportResolver } from 'app/shared/resolvers/report.resolver';
-import { ReportLevelMemoComponent } from '../../shared/report-level-memo/report-level-memo.component';
-import { PrintPreviewComponent } from 'app/reports/shared/print-preview/print-preview.component';
-import { ReportIsEditableGuard } from '../../../shared/guards/report-is-editable.guard';
-import { SubmitReportComponent } from '../../submission-workflow/submit-report.component';
-import { Report } from 'app/shared/models/reports/report.model';
-import { SubmitReportStatusComponent } from '../../submission-workflow/submit-report-status.component';
-import { CreateF3Step1Component } from '../create-workflow/create-f3-step1.component';
-import { ReportDetailedSummaryComponent } from './report-detailed-summary/report-detailed-summary.component';
-import { ReportSidebarSection } from 'app/layout/sidebar/menu-info';
 import { ReportSummaryComponent } from './report-summary/report-summary.component';
-import { Form3Service } from 'app/shared/services/form-3.service';
+import { ReportDetailedSummaryComponent } from './report-detailed-summary/report-detailed-summary.component';
+import { ReportResolver } from 'app/shared/resolvers/report.resolver';
+import { PrintPreviewComponent } from 'app/reports/shared/print-preview/print-preview.component';
+import { Report } from 'app/shared/models/reports/report.model';
+import { ReportSidebarSection } from 'app/layout/sidebar/menu-info';
+import { ReportLevelMemoComponent } from 'app/reports/shared/report-level-memo/report-level-memo.component';
+import { SubmitReportStatusComponent } from 'app/reports/submission-workflow/submit-report-status.component';
+import { SubmitReportComponent } from 'app/reports/submission-workflow/submit-report.component';
+import { ReportIsEditableGuard } from 'app/shared/guards/report-is-editable.guard';
+import { CreateSharedF3Component } from '../create-shared-f3/create-shared-f3.component';
+import { Form3XService } from 'app/shared/services/form-3x.service';
 import { FORM_3_SERVICE } from 'app/shared/services/base-form-3.service';
 
 // ROUTING NOTE:
@@ -20,12 +20,12 @@ import { FORM_3_SERVICE } from 'app/shared/services/base-form-3.service';
 // 1) The component will pull the active report from the ngrx store and not the ActivatedRoute.snapshot.
 // 2) The ReportResolver should not be declared on routes with a ReportIsEditableGuard declared.
 
-export const F3_ROUTES: Route[] = [
+export const F3X_ROUTES: Route[] = [
   {
     path: 'create/step1',
     title: 'Create a report',
-    component: CreateF3Step1Component,
-    providers: [{ provide: FORM_3_SERVICE, useClass: Form3Service }],
+    component: CreateSharedF3Component,
+    providers: [{ provide: FORM_3_SERVICE, useClass: Form3XService }],
     runGuardsAndResolvers: 'always',
     data: {
       sidebar: null,
@@ -34,16 +34,16 @@ export const F3_ROUTES: Route[] = [
   {
     path: 'create/step1/:reportId',
     title: 'Create a report',
-    component: CreateF3Step1Component,
-    providers: [{ provide: FORM_3_SERVICE, useClass: Form3Service }],
+    component: CreateSharedF3Component,
+    providers: [{ provide: FORM_3_SERVICE, useClass: Form3XService }],
     canActivate: [ReportIsEditableGuard],
     runGuardsAndResolvers: 'always',
   },
   {
     path: 'edit/:reportId',
     title: 'Edit a report',
-    component: CreateF3Step1Component,
-    providers: [{ provide: FORM_3_SERVICE, useClass: Form3Service }],
+    component: CreateSharedF3Component,
+    providers: [{ provide: FORM_3_SERVICE, useClass: Form3XService }],
     resolve: { report: ReportResolver },
     data: { sidebarSection: ReportSidebarSection.CREATE },
     runGuardsAndResolvers: 'always',
@@ -71,8 +71,8 @@ export const F3_ROUTES: Route[] = [
     resolve: { report: ReportResolver },
     data: {
       sidebarSection: ReportSidebarSection.REVIEW,
-      getBackUrl: (report?: Report) => '/reports/f3/detailed-summary/' + report?.id,
-      getContinueUrl: (report?: Report) => '/reports/f3/submit/' + report?.id,
+      getBackUrl: (report?: Report) => '/reports/f3x/detailed-summary/' + report?.id,
+      getContinueUrl: (report?: Report) => '/reports/f3x/submit/' + report?.id,
     },
     runGuardsAndResolvers: 'always',
   },
@@ -84,7 +84,7 @@ export const F3_ROUTES: Route[] = [
     resolve: { report: ReportResolver },
     data: {
       sidebarSection: ReportSidebarSection.REVIEW,
-      getNextUrl: (report?: Report) => '/reports/f3/submit/' + report?.id,
+      getNextUrl: (report?: Report) => '/reports/f3x/submit/' + report?.id,
     },
     runGuardsAndResolvers: 'always',
   },
@@ -96,8 +96,8 @@ export const F3_ROUTES: Route[] = [
     resolve: { report: ReportResolver },
     data: {
       sidebarSection: ReportSidebarSection.SUBMISSION,
-      getBackUrl: (report?: Report) => '/reports/f3/memo/' + report?.id,
-      getContinueUrl: (report?: Report) => '/reports/f3/submit/status/' + report?.id,
+      getBackUrl: (report?: Report) => '/reports/f3x/memo/' + report?.id,
+      getContinueUrl: (report?: Report) => '/reports/f3x/submit/status/' + report?.id,
     },
     runGuardsAndResolvers: 'always',
   },
