+++ front-end/src/app/reports/routes.ts
@@ -27,12 +27,12 @@ export const REPORTS_ROUTES: Route[] = [
   {
     path: 'f3',
     providers: [{ provide: ReportService, useClass: Form3Service }],
-    loadChildren: () => import('./base-f3/f3/routes').then((module) => module.F3_ROUTES),
+    loadChildren: () => import('./f3-shared/f3/routes').then((module) => module.F3_ROUTES),
   },
   {
     path: 'f3x',
     providers: [{ provide: ReportService, useClass: Form3XService }],
-    loadChildren: () => import('./base-f3/f3x/routes').then((module) => module.F3X_ROUTES),
+    loadChildren: () => import('./f3-shared/f3x/routes').then((module) => module.F3X_ROUTES),
   },
   {
     path: 'f99',
