+++ front-end/src/app/reports/submission-workflow/submit-report.component.spec.ts
@@ -16,7 +16,7 @@ import { DividerModule } from 'primeng/divider';
 import { RadioButtonModule } from 'primeng/radiobutton';
 import { BehaviorSubject } from 'rxjs';
 import { SubmitReportComponent } from './submit-report.component';
-import { F3X_ROUTES } from '../base-f3/f3x/routes';
+import { F3X_ROUTES } from '../f3-shared/f3x/routes';
 
 const routeDataSubject = new BehaviorSubject<{
   report: Report;
